+++ anvil_consortium_manager/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.15dev4"
+__version__ = "0.15dev5"